refactor: 사진 등록 가이드 섹션 컴포넌트화
