+++ src/features/upload/page.tsx
@@ -6,15 +6,11 @@ import BottomCTA from '@/components/ui/bottom-cta';
 import { API_PATH } from '@/lib/path';
 import { SourceType } from '@/types/common';
 
+import UploadGuide from './components/UploadGuide';
+
 interface UploadContainerProps {
   source: SourceType;
 }
-
-const guideImages = [
-  '/images/upload/upload-guide-1.svg',
-  '/images/upload/upload-guide-2.svg',
-  '/images/upload/upload-guide-3.svg',
-];
 
 // TODO: API 요청 시 헤더 또는 바디에 source 포함시키기
 const UploadContainer = (_props: UploadContainerProps) => {
@@ -162,7 +158,7 @@ const UploadContainer = (_props: UploadContainerProps) => {
                 />
                 <div className='absolute inset-0 flex flex-col items-center justify-center'>
                   <img
-                    src='/images/upload/icon-hammer.svg'
+                    src='/images/common/icon-hammer.svg'
                     alt=''
                     aria-hidden='true'
                   />
@@ -187,41 +183,7 @@ const UploadContainer = (_props: UploadContainerProps) => {
         </section>
 
         {/* 사진 가이드 섹션 */}
-        <section className='flex w-full flex-col items-center gap-y-5 bg-[#fff6ed] px-5 pt-10 pb-7.5'>
-          <p className='text-18-semibold'>어떤 사진을 올려야하나요?</p>
-
-          <div className='flex w-full flex-col items-center gap-y-2'>
-            <div className='flex items-center gap-x-2'>
-              <img
-                src='/images/upload/check.svg'
-                alt=''
-                aria-hidden='true'
-              />
-              <p>공간, 통로, 좌석이 함께 보이는 사진</p>
-            </div>
-
-            <div className='flex items-center gap-x-2'>
-              <img
-                src='/images/upload/x.svg'
-                alt=''
-                aria-hidden='true'
-              />
-              <p>사람, 음식, 천장 등 클로즈업 사진</p>
-            </div>
-          </div>
-
-          {/* 업로드 가이드 이미지 */}
-          <div className='flex min-h-[84px] w-full gap-x-0.5'>
-            {guideImages.map((src, index) => (
-              <img
-                key={src}
-                src={src}
-                alt={`가이드 이미지 ${index + 1}`}
-                className='h-full flex-1 object-contain'
-              />
-            ))}
-          </div>
-        </section>
+        <UploadGuide />
       </main>
 
       {/* 하단 CTA 버튼 */}
